+++ src/app/alignments/alignments.module.ts
@@ -4,8 +4,9 @@ import { FormsModule } from '@angular/forms';
 import { TranslateModule } from '@ngx-translate/core';
 import { GraphModule } from '../graph/graph.module';
 import { WidgetModule } from '../widget/widget.module';
-import { AlignmentsListPanelComponent } from './alignments-list-panel/alignments-list-panel.component';
-import { AlignmentsListComponent } from './alignments-list-panel/alignments-list.component';
+import { AlignmentTreeNodeComponent } from './alignments-tree-panel/alignment-tree-node.component';
+import { AlignmentsTreePanelComponent } from './alignments-tree-panel/alignments-tree-panel.component';
+import { AlignmentsTreeComponent } from './alignments-tree-panel/alignments-tree.component';
 import { AlignmentsTableComponent } from './alignments-table/alignments-table.component';
 import { AlignmentsView } from './alignments-view.component';
 import { AlignmentsComponent } from './alignments.component';
@@ -15,10 +16,11 @@ import { AlignmentsModalsServices } from './modals/alignments-modal.service';
 @NgModule({
     declarations: [
         AlignmentsComponent,
-        AlignmentsListComponent,
         AlignmentsModal,
-        AlignmentsListPanelComponent,
+        AlignmentsTreePanelComponent,
         AlignmentsTableComponent,
+        AlignmentsTreeComponent,
+        AlignmentTreeNodeComponent,
         AlignmentsView
     ],
     imports: [
@@ -29,7 +31,7 @@ import { AlignmentsModalsServices } from './modals/alignments-modal.service';
         WidgetModule
     ],
     exports: [
-        AlignmentsListPanelComponent,
+        AlignmentsTreePanelComponent,
         AlignmentsView
     ],
     providers: [