+++ src/app/models/Alignments.ts
@@ -1,4 +1,0 @@
-export enum AlignmentContext {
-    local = "local", //(in tabset) source project is the one in used, target project must be selected
-    global = "global" //(dedicated page from navbar) both source and target projects must be selected
-}